refactor: Update Account Settings Page and Optimize Server Rendering
